extract .invite-code div into RoomCode component
