+++ src/components/common/RoomCode/index.ts
@@ -1,0 +1,1 @@
+export { default as RoomCode } from './RoomCode';
